Update pyproject.toml
An exact version pin is problematic, since it means that you don't get updates and improvements. Hamilton follows semantic versioning, so a ^1.83.2 will mean that anything greater but less than 2.0.0 is good. 

This will help us on the Hamilton side too because we'll have most users on the latest version, rather than pinned at specific old points in time.